requirements.txt
Django==4.2
djangorestframework==3.14.0
transformers==4.31.0
torch==2.5.0
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
langdetect==1.0.9